rds/client/python/requirements.txt
psycopg2-binary==2.8.6
pymssql==2.2.2